Add text shadow and rotate-by effect support
Introduces shadow_color and shadow_offset properties to the Text class, enabling text shadow rendering. Adds fx.rotate_by effect for relative rotation animations and updates fx.rotate_to effect to support rotation direction. Also improves Transform and effect parameter types for consistency, and enhances renderer logic for size handling.

+++ include/Math.hpp
@@ -116,10 +116,10 @@ Vec2 operator*(double lhs, const Vec2& rhs);
 
 struct Transform
 {
-    Vec2 pos{0.0f, 0.0f};
+    Vec2 pos{0.0, 0.0};
     Vec2 size{};          // Explicit size (empty = use texture/srcRect size)
-    float angle = 0.0f;   // In radians
-    Vec2 scale{1.0f, 1.0f};
+    double angle = 0.0;   // In radians
+    Vec2 scale{1.0, 1.0};
     Anchor anchor = Anchor::TopLeft;
     Vec2 pivot{0.5, 0.5}; // Normalized pivot point, centered by default
 };
